refactor(graphql): rename filters to use snake case naming
